requirements.txt
Django==1.8.7
djangorestframework==3.3.1
dj-database-url==0.3.0
